src/novo_backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
fastapi[all]==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.3
portalocker==2.8.2
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.3
sniffio==1.3.1
starlette==0.36.3
tinydb==4.8.0
typing-extensions==4.10.0
opencv-python==4.9.0.80
uvicorn==0.28.1
qreader==3.12
pyserial==3.5
yaspin==3.0.1
rich==13.7.1